Fixes #11579 - Reports show/destroy restricted by host authorization (CVE-2015-5233)
ReportsController 'show' and 'destroy' now perform a check to see if
the User is authorized to see the Host associated with the Report. In
case it's not, it returns 404, as to not give hints whether a Report
ID or Host ID are valid.

I followed the same approach on the API controllers. 'last' was not
vulnerable due to using my_reports which performs the necessary
check on 'view_hosts' permission.

(cherry picked from commit 293036dfa71ae70624663647f1ef70798bf53d3e)

Conflicts:
	app/controllers/api/v1/reports_controller.rb
	app/controllers/api/v2/reports_controller.rb
	app/controllers/reports_controller.rb
	test/functional/api/v1/reports_controller_test.rb
	test/functional/api/v2/reports_controller_test.rb